+++ custom_crm/jaybutani_crm_customisation/changelog.md
@@ -1,5 +1,10 @@
 ## [18.0.1.0.0] - 2025-04-10 | Propagate Opportunity Name to Downstream Objects
 
-- inherit sale.order, stock.picking, project.project, mrp.production and add job_name field
-- add functionality of sale oredr job name is consider in all above model job name field
+- inherit sale_order, stock_picking, project_project, mrp_production and add job_name field
+- add functionality of sale order job name is consider in all above model job name field
 - add functionality of editable or non-editable
+
+## [18.0.1.0.1] - 2025-04-14 | Propagate Opportunity Name to Downstream Objects
+
+- changes on sale order instead of onchange use compute method
+- add doc string and formatting